Experimental

+++ pdf/__init__.py
@@ -18,16 +18,16 @@ def generate_pdf(book_path, base_filename, tex_file):
     proc = subprocess.Popen(cmd)
     proc.communicate()
 
-    os.unlink(filepath(book_path, base_filename, 'toc'))
-    os.unlink(filepath(book_path, base_filename, 'log'))
-    os.unlink(filepath(book_path, base_filename, 'aux'))
-
     retcode = proc.returncode
 
     if not retcode == 0:
         os.unlink(pdf_file)
         raise ValueError('Error {} executing command: {}'.format(retcode, ' '.join(cmd)))
     else:
+        if os.path.isfile(filepath(book_path, base_filename, 'toc')):
+            os.unlink(filepath(book_path, base_filename, 'toc'))
+        os.unlink(filepath(book_path, base_filename, 'log'))
+        os.unlink(filepath(book_path, base_filename, 'aux'))
         show_file(pdf_file)
 
     return pdf_file
